port: fix LP64 base-type widening (LONG/ULONG were 64-bit on Linux/macOS)
Windows types are LLP64: LONG/ULONG are always 32-bit. The POSIX shim
typedef'd them to C long / unsigned long, which are 64-bit under LP64
(Linux/macOS). That silently doubled every struct using them - GOFFSET
{LONG du,dv} became 16 bytes vs the Rust ffi.rs mirror's 8, so
ScriptPlaceOpenType wrote pGoffset[i] at 16-byte strides into the 7-entry
(56-byte) Rust buffer, overflowing after the 4th element (i=3 dv @ +56).
Windows/MinGW (LLP64, long=32) was unaffected, which is why the crash was
Linux-only - surfaced by ASan as a heap-buffer-overflow right after a
correct PARITY_MATCH (the shape trace never reads place offsets, so the
glyph parity was already equal). Fix LONG=int, ULONG=unsigned int and make
the winbase.h Interlocked*/CRITICAL_SECTION use LONG so volatile LONG*
callers (trace.c) still type-check.

+++ tools/wine_usp/port/posix/windef.h
@@ -90,13 +90,18 @@
 #define VOID void
 
 /* ---- base scalars ---- */
+/* IMPORTANT: Windows types are LLP64 — LONG/ULONG are ALWAYS 32-bit even on
+ * 64-bit Windows. On Linux/macOS (LP64) the C `long` is 64-bit, so typedefing
+ * LONG to `long` would silently double every struct that uses it (GOFFSET,
+ * ABC, HRESULT, RECT, TEXTMETRICW, ...) and desync the C ABI from the Rust
+ * ffi.rs mirrors (which use i32/u32). Hence LONG=int, ULONG=unsigned int. */
 typedef char            CHAR;
 typedef short           SHORT;
-typedef long            LONG;
+typedef int             LONG;
 typedef int             INT;
 typedef unsigned char   UCHAR;
 typedef unsigned short  USHORT;
-typedef unsigned long   ULONG;
+typedef unsigned int    ULONG;
 typedef unsigned int    UINT;
 typedef float           FLOAT;
 typedef double          DOUBLE;
